requirements.txt
Werkzeug==3.0.1
ipykernel==6.29.3
ipython==8.18.1
bs4==0.0.2
beautifulsoup4==4.12.3
lxml==5.1.0
numpy==1.26.4
pandas==2.2.1
PyJWT==2.9.0
psycopg2-binary==2.9.9
selenium==4.18.1
python-dotenv==1.0.1
azure-storage-blob==12.24.0
pynput==1.7.6
openai==1.59.9